FullStop and Timer Adjustments
Fullstop is "in." I am not sure what else this would do for now. I will say we probably need some visual effect for it, but the actual ability itself is implemented. I also documented timer and adjusted it so that its more readable and modifiable. I really love our timer right now.